% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_a_package.R
\name{add.a.package}
\alias{add.a.package}
\title{Discover new packages, related to the ones you have}
\usage{
add.a.package(relationship = "suggests",
  my_packs = rownames(installed.packages()), kind.of = "to_packages",
  neighbors.neighbors = FALSE, nAdd = 1)
}
\arguments{
\item{relationship}{The kind of relationship you use to recommend. It can be 'suggests' or 'depends'.}

\item{my_packs}{The packages that you are interested in to use as reference. Default is your installed packages.}

\item{kind.of}{The direction of the relations. If you want packages recommending my_packs, it should be "to_packages", and if you want packages recommended by my_packs, it should be "from_packages". It can also be "all".}

\item{neighbors.neighbors}{If it is TRUE, only the neighbors of the new package are included in the output.}
}
\value{
An igraph object.
}
\description{
This function gives you an igraph network with your packages and the first one that recommend.me gives you.
}
\examples{
g = add.a.package(gs)
plot(g)
}
