% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplification_coef.R
\name{amplification.coefficient}
\alias{amplification.coefficient}
\title{Discover new packages, related to the ones you have}
\usage{
amplification.coefficient(pack, gs, gd,
  internals = rownames(installed.packages()))
}
\arguments{
\item{pack}{The package of interest}

\item{gs}{The package graph, as an igraph object. Usually the Suggests graph.}

\item{gd}{The package graph, as an igraph object. Usually the Depends graph.}

\item{internals}{The set of packages to be considered. By default, your packages.}
}
\value{
A data.frame with the different coefficients.
}
\description{
A coefficient that measures how much your network horizons will expand after adding a package.
}
\examples{
pack = 'fuzzywuzzyR'
amplification.coefficient(pack,gs,gd)
}
