% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cats}
\alias{cats}
\title{A data frame with the categories of the packages}
\format{An object of class igraph.}
\usage{
data(cats)
}
\description{
A data frame with the categories of the packages
}
\examples{
data(cats)
print(cats)
}
\keyword{datasets}
