% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{desc}
\alias{desc}
\title{A data frame with the descriptions of the packages}
\format{An igraph class object.}
\usage{
data(des)
}
\description{
A data frame with the descriptions of the packages
}
\examples{
data(des)
print(desc)
}
\keyword{datasets}
