% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\title{Auxiliary functions for the leaflet plotting}
\usage{
description(pack)
}
\arguments{
\item{pack}{The package of interest}

\item{graph}{The graph to be plotted.}

\item{layout}{The graph layout}

\item{texto}{The text to be parsed.}

\item{nwords}{The number of words per line.}
}
\value{
Multiple values depending on the function
}
\description{
This functions help to manage the text in the leaflet plotting
}
