% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{downloads}
\alias{downloads}
\title{A data frame with the number of packages downloads}
\format{An object of class igraph.}
\usage{
data(downloads)
}
\description{
A data frame with the number of packages downloads
}
\examples{
data(downloads)
print(dowloads)
}
\keyword{datasets}
