% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gd}
\alias{gd}
\title{Depends graph}
\format{An igraph class object.}
\usage{
data(dependsGraph)
}
\description{
Package graph, with packages conected by Depends relations
}
\examples{
data(dependsGraph)
plot(gd)
}
\keyword{datasets}
