% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ge}
\alias{ge}
\title{Enhances graph}
\format{An igraph class object.}
\usage{
data(enhancesGraph)
}
\description{
Package graph, with packages conected by Enhances relations
}
\examples{
data(enhancesGraph)
plot(ge)
}
\keyword{datasets}
