% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_labels.R
\name{generate_labels}
\alias{generate_labels}
\title{Auxiliary functions for the leaflet plotting}
\usage{
generate_labels(graph, nwords = 15)
}
\arguments{
\item{graph}{The graph to be plotted.}

\item{nwords}{The number of words per line.}
}
\value{
Multiple values depending on function
}
\description{
This functions help manage the text in the leaflet plotting
}
