% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gi}
\alias{gi}
\title{Imports graph}
\format{An igraph class object .}
\usage{
data(importsGraph)
}
\description{
Package graph, with packages conected by Imports relations
}
\examples{
data(importsGraph)
plot(gi)
}
\keyword{datasets}
