% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_me_one.R
\name{give.me.one}
\alias{give.me.one}
\title{Discover new packages, related to the ones you have}
\usage{
give.me.one(relationship = "suggests",
  my_packs = rownames(installed.packages()), kind.of = "to_packages",
  deterministic = F, apply.degree.filter = F)
}
\arguments{
\item{relationship}{The kind of relationship to use for recommendation. It can be 'suggests' or 'depends'.}

\item{my_packs}{The packages that you are interested to use as reference. Default is your installed packages}

\item{kind.of}{The direction of the relations. If you want packages recommending my_packs, it should be "to_packages", and if you want packages recommended by my_packs, it should be "from_packages". It can also be "all".}

\item{deterministic}{If TRUE, the function returns the first one of the list generated from recommend.me. If FALSE, it samples from the list generated from recommend.me, using the scores as sampling probabilities.}

\item{apply.degree.filter}{If TRUE, the score of each package is multiplied by the fraction of its edges that conect to my_packs.}
}
\value{
A recommended package.
}
\description{
Shortcome for recommend.me. Its a good choice from the list it provides.
}
\examples{
pack = give.me.one(gs)
install.packages(pack)
}
