% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2graph.R
\name{graph2graph}
\alias{graph2graph}
\title{Get the equivalent network with other relations}
\usage{
graph2graph(relationship, sgs)
}
\arguments{
\item{relationship}{The new relation to be used. It can be 'suggests' or 'depends'.}

\item{sgs}{Your own package network}
}
\value{
An igraph object.
}
\description{
This function gives you the equivalent network in another set of relationships.
}
\examples{
sgd = graph2graph('depends',sgs)
plot(sgd)
}
