% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gs}
\alias{gs}
\title{Suggests graph}
\format{An igraph class object .}
\usage{
data(suggestsGraph)
}
\description{
Package graph, with packages conected by Suggests relations
}
\examples{
data(suggestsGraph)
plot(gs)
}
\keyword{datasets}
