% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_conections.R
\name{list_conections}
\alias{list_conections}
\title{Auxiliary functions for the leaflet plotting}
\usage{
list_conections(graph, layout)
}
\arguments{
\item{graph}{The graph to be plotted.}

\item{layout}{The layout of the graph}

\item{des}{The descriptions dataframe.}

\item{cats}{The categories dataframe.}

\item{texto}{The text to be parsed.}

\item{nwords}{The number of words per line.}

\item{pack}{The package of interest}
}
\value{
Multiple values depending on function
}
\description{
This functions help manage the text in the leaflet plotting
}
