% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_connections.R
\name{list_connections}
\alias{list_connections}
\title{Auxiliary functions for the leaflet plotting}
\usage{
list_connections(graph, layout)
}
\arguments{
\item{graph}{The graph to be plotted.}

\item{layout}{The layout of the graph}
}
\value{
Multiple values depending on function
}
\description{
This functions help manage the text in the leaflet plotting
}
