% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_network.R
\name{my_network}
\alias{my_network}
\title{Your own R-Network of packages, as an igraph object}
\usage{
my_network()
}
\value{
A list with igraph objects, one for each input graph.
}
\description{
This function gives you your own package network, based on the avaiable general networks.
}
\examples{
g_list = my_network()
plot(g_list[['Suggests']])
}
