% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_packages.R
\name{my_packages}
\alias{my_packages}
\title{Your own packages}
\usage{
my_packages(file = NULL)
}
\arguments{
\item{file}{The file path to write down your packages}
}
\value{
A string with the names of your installed packages
}
\description{
This function prints your packages on screen or in a file.
}
\examples{
my_packages()
}
