% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_step2my_net.R
\name{one_step2my_net}
\alias{one_step2my_net}
\title{Your own R-Network, plus its first neighbors}
\usage{
one_step2my_net(kind.of = "all")
}
\arguments{
\item{kind.of}{The direction considered for the neighborhood. If you want packages pointing to yours, use 'to_packages', if you want packages pointed by yours, use 'from_packages'. If you want them all, use 'all'}
}
\value{
A list with igraph objects, one for each input graph.
}
\description{
This function gives you your own package network, based on the avaiable general networks, plus its first neighbots
}
\examples{
g = one_step2my_net(gd,gs)$Sug
plot(g)
}
