% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neighbors.R
\name{plot_neighbors}
\alias{plot_neighbors}
\title{Discover new packages, related to the ones you have}
\usage{
plot_neighbors(relationship = "suggests",
  my_packs = rownames(installed.packages()), kind.of = "to_packages",
  order = 1, point.size = "score", apply.degree.filter = F,
  nwords = 5)
}
\arguments{
\item{relationship}{The kind of relationship to use for neighbouring. It can be 'suggests' or 'depends'.}

\item{my_packs}{The packages that you are interested to use as reference. Default is your installed packages}

\item{kind.of}{The direction of the relations. If you want packages recommending my_packs, it should be "to_packages", and if you want packages recommended by my_packs, it should be "from_packages". It can also be "all".}

\item{order}{The order of the neighborhood. It has to be equal or bigger than zero.}

\item{point.size}{The aspect of the packages giving their size. If 'score', their size is given based in their score. If 'downloads', their size is base in ther daily downloads. Else its the same for all of them.}

\item{apply.degree.filter}{If TRUE, the score of each package is multiplied by the fraction of its edges that conect to my_packs.}

\item{nwords}{The number of words appearing in the description of the package, per row.}
}
\value{
A list with packages and scores for each one.
}
\description{
Plots your network, plus its firts neighbors, using leaflet.
}
\examples{
plot_neighbors(gs,point.size='score',downloads=downloads)
}
