% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proposal.R
\name{plot_proposal}
\alias{plot_proposal}
\title{Discover new packages, related to the ones you have}
\usage{
plot_proposal(relationship = "suggests",
  my_packs = rownames(installed.packages()), vertex.label.dist = -1,
  vertex.label.degree = pi/2, vertex.size = 3, arrow.size = 0,
  kind.of = "to_packages", new.label = FALSE,
  neighbors.label = FALSE, others.label = FALSE,
  neighbors.neighbors.label = FALSE, only.neighbors = FALSE,
  neighbors.neighbors = FALSE, label.cex = 1, label.color = "black",
  layout = NULL, on.leaflet = FALSE, nAdd = 1,
  recenter.suggested = !on.leaflet, nwords = 5,
  point.size = "downloads")
}
\arguments{
\item{relationship}{The kind of relationship to use for recommendation. It can be 'suggests' or 'depends'.}

\item{my_packs}{The packages that you are interested to use as reference. Default is your installed packages}

\item{kind.of}{The direction of the relations. If you want packages recommending my_packs, it should be "to_packages", and if you want packages recommended by my_packs, it should be "from_packages". It can also be "all".}

\item{neighbors.neighbors}{If it is TRUE, only the neighbors of the new package are considered.}

\item{on.leaflet}{The plot should be done on leaflet?}

\item{point.size}{The aspect of the packages giving their size. If 'score', their size is given based in their score. If 'downloads', their size is base in ther daily downloads. Else its the same for all of them.}

\item{...}{Igraph plotting parameters}
}
\value{
A list with packages and scores for each one.
}
\description{
Plots your network plus a recommended package added.
}
\examples{
plot_proposal(gs,on.leaflet=T,downloads=downloads)
}
