% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_meOLD.R
\name{recommend.meOLD}
\alias{recommend.meOLD}
\title{Discover new packages, related to the ones you hace}
\usage{
recommend.meOLD(gs, my_packs = rownames(installed.packages()),
  kind.of = "to_packages")
}
\arguments{
\item{gs}{The package graph, as an igraph object. Usually the Suggests graph.}

\item{my_packs}{The packages that you are interested to use as reference. Default is your installed packages}

\item{kind.of}{The direction of the relations. If you want packages recommending my_packs, it should be "to_packages", and if you want packages recommended by my_packs, it should be "from_packages". It can also be "all".}
}
\value{
A list with packages and scores for each one.
}
\description{
This function provides you with recommendations of new packages to install.
}
\examples{
pack_list = recommend.me(graph.dep,graph.sug)
install.packages(names(pack_list)[1])
}
