% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_me.R
\name{recommend_me}
\alias{recommend_me}
\title{Rank available packages to be installed}
\usage{
recommend_me(relationship = "suggests",
  my_packs = rownames(utils::installed.packages()),
  kind.of = "to_packages", apply.degree.filter = F, niter = 50,
  nMax = NULL)
}
\arguments{
\item{relationship}{The kind of relationship you use to recommend. It can be 'suggests','imports','enhances' or 'depends'.}

\item{my_packs}{The packages that you are interested to use as reference. Default is your installed packages}

\item{kind.of}{The direction of the relations. If you want packages recommending my_packs, it should be "to_packages", and if you want packages recommended by my_packs, it should be "from_packages". It can also be "all".}

\item{apply.degree.filter}{If TRUE, the score of each package is multiplied by the fraction of its edges that conect to my_packs.}

\item{niter}{The number of interations done to calculate the score.}

\item{nMax}{The max number of elements to be retrieved. Default is the full list.}
}
\value{
A list with packages and scores for each one.
}
\description{
This function provides you with recommendations of new packages to install.
}
\examples{
\dontrun{
pack_list = recommend_me()
install.packages(names(pack_list)[1])
}
}
