% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_your_packages.R
\name{send_your_packages}
\alias{send_your_packages}
\title{Provide your packages to the authors, for further research}
\usage{
send_your_packages()
}
\value{
Nothing, it just sends the e-mail.
}
\description{
This function sends your packages to the authors, using the 'mailR' package.
}
\examples{
\dontrun{
send_your_packages()
}
}
