% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_are_you.R
\name{who.are.you}
\alias{who.are.you}
\title{Discover new packages, related to the ones you have}
\usage{
who.are.you(pack, plot.it = TRUE, facts = TRUE, on.leaflet = FALSE,
  nwords = 5, add_my_packs = FALSE, my_packs_neighbors = TRUE,
  point.size = "downloads", ...)
}
\arguments{
\item{pack}{The package(s) of interest.}

\item{plot.it}{Should the network be plotted? Default is TRUE}

\item{facts}{Should the output contain facts about the package? Default is TRUE}

\item{on.leaflet}{Should the plot be done in leaflet? Default is FALSE}

\item{add_my_packs}{Should your packages be added to the plot? Default is FALSE}

\item{my_packs_neighbors}{If add_my_packs is TRUE, should only keep the surrounders of pack? Default is TRUE}

\item{point.size}{The aspect of the packages giving their size. If 'score', their size is given based in their score. If 'downloads', their size is base in ther daily downloads. Else its the same for all of them.}

\item{gs}{The package graph, as an igraph object. Usually the Suggests graph.}

\item{gd}{The package graph, as an igraph object. Usually the Depends graph.}
}
\value{
A dataframe and also a plot.
}
\description{
This function gives you an ego network of a package (the package and its surrounders).
}
\examples{
who.are.you('ggplot2',gs,gd,des=desc,cats=cats,downloads=downloads)
}
