% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_string.R
\name{cut_string}
\alias{cut_string}
\title{Auxiliary functions for the leaflet plotting}
\usage{
cut_string(texto, nwords = 15)
}
\arguments{
\item{texto}{The text to be parsed.}

\item{nwords}{The number of words per line.}

\item{graph}{The graph to be plotted.}

\item{layout}{The layout of the graph}

\item{pack}{The package of interest}
}
\value{
Multiple values depending on the function
}
\description{
These functions help to manage the text in the leaflet plotting
}
