% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{description}
\alias{description}
\title{Auxiliary functions for the leaflet plotting}
\usage{
description(pack)
}
\arguments{
\item{pack}{The package of interest}
}
\value{
Multiple values depending on the function
}
\description{
This functions help to manage the text in the leaflet plotting
}
\examples{
description('leaflet')
}
