% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expre2graph.R
\name{expre2graph}
\alias{expre2graph}
\title{Discover new packages, related to the ones you have}
\usage{
expre2graph(expression, plot.it = FALSE, first.neighbors = FALSE,
  generate_output = TRUE, point.size = "downloads",
  min.point.size = 15, max.point.size = 30, nwords = 5)
}
\arguments{
\item{expression}{The expression of interest}

\item{plot.it}{Should the graph be plotted? Default is FALSE}

\item{first.neighbors}{Should the first neighbors be added to the graph? Default is FALSE}

\item{generate_output}{Should graph objects be retrieved? Default is TRUE}

\item{point.size}{The packages attribute which defines their size. If 'score', their size is defines based on their score. If 'downloads', their size is based on their daily downloads. Otherwise it is the same size for all of them.}

\item{min.point.size}{The min point size, passed to leaflet. Default is 15}

\item{max.point.size}{The max point size, passed to leaflet. Default is 30}

\item{nwords}{Number of words per line in the description of the package}
}
\value{
Graph plot in leaflet, and an igraph object.
}
\description{
Obtain a graph induced by the packages containing certain expression in their description. The expression is matched using agrep.
}
\examples{
expression = 'information'
expre2graph(expression)
}
