% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neighbors.R
\name{plot_neighbors}
\alias{plot_neighbors}
\title{Find nearest packages to a selected group}
\usage{
plot_neighbors(relationship = "suggests",
  my_packs = rownames(utils::installed.packages()),
  kind.of = "to_packages", order = 1, point.size = "score",
  return.map = T, plot.it = F, apply.degree.filter = F, nwords = 5,
  min.point.size = 15, max.point.size = 30)
}
\arguments{
\item{relationship}{The kind of relationship you use to recommend. It can be 'suggests','imports','enhances' or 'depends'.}

\item{my_packs}{The packages that you are interested to use as reference. Default is your installed packages}

\item{kind.of}{The direction of the relations. If you want packages recommending my_packs, it should be "to_packages", and if you want packages recommended by my_packs, it should be "from_packages". It can also be "all".}

\item{order}{The order of the neighborhood. It has to be equal or bigger than zero.}

\item{point.size}{The aspect of the packages giving their size. If 'score', their size is given based in their score. If 'downloads', their size is base in ther daily downloads. Else its the same for all of them.}

\item{return.map}{Should the leaflet and igraph object be returned?}

\item{plot.it}{Should the leaflet object be printed?}

\item{apply.degree.filter}{If TRUE, the score of each package is multiplied by the fraction of its edges that conect to my_packs.}

\item{nwords}{The number of words appearing in the description of the package, per row.}

\item{min.point.size}{The min point size, passed to leaflet. Default is 15}

\item{max.point.size}{The max point size, passed to leaflet. Default is 30}
}
\value{
A list with packages and scores for each one.
}
\description{
Plot your network, plus its firts neighbors, using leaflet.
}
\examples{
plot_neighbors()
}
