% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_are_you.R
\name{who.are.you}
\alias{who.are.you}
\title{Discover new packages, related to the ones you have}
\usage{
who.are.you(pack, plot.it = TRUE, nwords = 5, add_my_packs = FALSE,
  my_packs_neighbors = FALSE, point.size = "downloads",
  min.point.size = 15, max.point.size = 30)
}
\arguments{
\item{pack}{The package of interest.}

\item{plot.it}{Should the network be plotted? Default is TRUE}

\item{nwords}{The number of words per line in the package's description}

\item{add_my_packs}{Should your packages be added to the plot? Default is FALSE}

\item{my_packs_neighbors}{If add_my_packs is TRUE, should only keep the surrounders of pack? Default is FALSE}

\item{point.size}{The aspect of the packages giving their size. If 'score', their size is given based in their score. If 'downloads', their size is base in ther daily downloads. Else its the same for all of them. If you want to specify a size use min.point.size}

\item{min.point.size}{The min point size, passed to leaflet. Default is 15}

\item{max.point.size}{The max point size, passed to leaflet. Default is 30}
}
\value{
A dataframe and also a plot.
}
\description{
This function gives you an ego network of a package (the package and its surrounders).
}
\examples{
who.are.you('leaflet')
}
